Reestructura proyecto
