+++ src/core/database/prisma.module.ts
@@ -1,5 +1,5 @@
 import { Global, Module } from '@nestjs/common';
-import { PrismaService } from 'src/infrastructure/services/prisma.service';
+import { PrismaService } from 'src/core/database/prisma.service';
 
 @Module({
   providers: [PrismaService],
