+++ src/modules/auth/infrastructure/controllers/app.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from 'src/infrastructure/services/app.service';
+import { AppService } from 'src/app.service';
 
 
 @Controller()
