+++ src/modules/users/application/dto/registration-user.dto.ts
